internal/testutil: implement BitGen format
The BitGen format is intended to be a human-readable format that can
be converted into a bit-stream. It is intended to be used to generate
test files for various compression formats. The goal of BitGen is to
provide a way for the human author to encode documentation about the
specifics of a bit-stream (e.g. why this stream is bad, what edge
cases it tests, etc).