internal/prefix: report correct Buffered
For bytesReader and stringsReader, Buffered reported the underlying
buffer size and not the number of valid bytes buffered.

This is not necessarily a correctness issue since the underlying
bytes or string of the Reader can technically be seen as being
"buffered", but there is a performance penalty to reporting as such.