doc: update Brotli framing format to support appending
* Remove use FrameFooter and merge that functionality into the FrameBlock.
This allows the framing format to better support appending of new data.
Although a new FrameHeader cannot be defined, it allows appending to be
possible since termination of the format is determined by the fact that
EOF is hit after reading a FrameBlock.
* Also remove the use of single-byte values like 00 and 01 to identify
the type of field when reading in the forward direction. Instead, we
restrict NumRecords to be non-zero to ensure that it never conflicts
with the first byte of the SyncMarker.