flate: add zcheck flag to allow verification of TestReader vectors
The use of the bitgen format does not make it obvious whether the
resulting stream is even valid or not. Thus, we add a test for the
unit test values themselves by making sure that a third-source
also validates the expected output. The third-party source used is
the C zlib library, accessed using the Python wrapper.