internal/testutil: add support for quoted strings in BitGen format
I have always wanted to use quoted strings in the BitGen format to
encode human readable strings in compression testdata. However, quoted
strings present a challenge to the parser since they may contain spaces.
This complicates the previous method which simply splitted on all spaces.

In this commit, we rewrite the parser to be able to handle quoted strings.
We heavily rely on fmt.Fscanf to do most of the processing for us.