flate: add test to ensure Reader never peeks more than necessary
Commit feffc6c3aeb4074599b191f935a7d2663709964d contained a change on prefix/reader.go:266:
	+ cntPeek := int(nb+(-nb&7)) / 8
	- cntPeek := 8 // Minimum Peek amount to make progress

This change ensures that we never call Peek for more bytes than is actually
necessary. However, the above commit never added a test to ensure that
this property holds true.

The fix was necessary to ensure that the Reader would never get blocked trying
to read bytes after the sync marker from a potentially blocking io.Reader
(e.g., a network socket). Instead, the Reader should report back to the user
with whatever data it was able to decompress so far.

This is a property that is guaranteed by the standard library's implementation
of flate and is relied upon by some networking protocols. We add this test here
to ensure that this property always holds true in this implementation as well.