bzip2: add support for reading multistream files
The BZip2 format allows back-to-back concatenation of files such that
the decompressed output of a series of BZip2 files is equivalent to
the concatenation of a series of uncompressed BZip2 files.

We alter TestRoundTrip to remove the canary byte check since the BZip2
format is "defined" to end when there are no more bytes to read and
we are right after a stream footer.