bzip2: disable checksum checking in gofuzz mode
The probability that a randomly fuzz-generated BZip2 file has the
proper checksums is very rare. This makes it very difficult for gofuzz
to properly explore the code and discover issues. Thus, we augment
the bzip2 code to disable checksums when running with gofuzz.

Since the C implementation still requires proper checksums, we use the
Go implementation to record the proper checksums and to apply them
into the original data stream before passing to the C decoder.