internal/testutil: add grouping to BitGen format
Certain tests would benefit from being able to repeat a sequence of tokens.
Thus, we add grouping to the BitGen format to allow grouping a series of
tokens within a set of "(" and ")". Similar to other tokens, these allow
appending a repeator decorator, which executes all tokens within the group
the specified number of times.